linux-dist: Move evdev/joystick code into separate files
