+++ core/linux-dist/main.cpp
@@ -33,17 +33,16 @@
   map<int, int> x11_keymap;
 #endif
 
-#if !defined(ANDROID) && HOST_OS != OS_DARWIN && !defined(TARGET_EMSCRIPTEN)
+#if defined(USES_HOMEDIR)
   #include <sys/stat.h>
-  #include <sys/types.h>
 #endif
 
 #if defined(USE_EVDEV)
-  #include "linux-dist/evdev_mappings.h"
+  #include "linux-dist/evdev.h"
 #endif
 
 #if defined(USE_JOYSTICK)
-  #include <linux/joystick.h>
+  #include "linux-dist/joystick.h"
 #endif
 
 #ifdef TARGET_PANDORA
@@ -91,45 +90,9 @@ u8 lt[4] = {0, 0, 0, 0};
 u32 vks[4];
 s8 joyx[4], joyy[4];
 
-enum DreamcastController
-{
-  DC_BTN_C       = 1,
-  DC_BTN_B       = 1<<1,
-  DC_BTN_A       = 1<<2,
-  DC_BTN_START   = 1<<3,
-  DC_DPAD_UP     = 1<<4,
-  DC_DPAD_DOWN   = 1<<5,
-  DC_DPAD_LEFT   = 1<<6,
-  DC_DPAD_RIGHT  = 1<<7,
-  DC_BTN_Z       = 1<<8,
-  DC_BTN_Y       = 1<<9,
-  DC_BTN_X       = 1<<10,
-  DC_BTN_D       = 1<<11,
-  DC_DPAD2_UP    = 1<<12,
-  DC_DPAD2_DOWN  = 1<<13,
-  DC_DPAD2_LEFT  = 1<<14,
-  DC_DPAD2_RIGHT = 1<<15,
-
-  DC_AXIS_LT = 0X10000,
-  DC_AXIS_RT = 0X10001,
-  DC_AXIS_X  = 0X20000,
-  DC_AXIS_Y  = 0X20001,
-};
-
 void emit_WriteCodeCache();
 
-/* evdev input */
-
 #if defined(USE_EVDEV)
-
-  struct s_controller
-  {
-    int fd;
-    ControllerMapping* mapping;
-  };
-
-  typedef struct s_controller Controller;
-
   /* evdev input */
   static Controller controllers[4] = {
     { -1, NULL },
@@ -137,362 +100,11 @@ void emit_WriteCodeCache();
     { -1, NULL },
     { -1, NULL }
   };
-
-  #define EVDEV_DEVICE_STRING "/dev/input/event%d"
-
-  #ifdef TARGET_PANDORA
-    #define EVDEV_DEFAULT_DEVICE_ID_1 4
-  #else
-    #define EVDEV_DEFAULT_DEVICE_ID_1 0
-  #endif
-
-  #define EVDEV_DEFAULT_DEVICE_ID(port) (port == 1 ? EVDEV_DEFAULT_DEVICE_ID_1 : -1)
-
-  int input_evdev_init(Controller* controller, const char* device)
-  {
-    char name[256] = "Unknown";
-
-    printf("evdev: Trying to open device at '%s'\n", device);
-
-    int fd = open(device, O_RDONLY);
-
-    if (fd >= 0)
-    {
-      fcntl(fd, F_SETFL, O_NONBLOCK);
-      if(ioctl(fd, EVIOCGNAME(sizeof(name)), name) < 0)
-      {
-        perror("evdev: ioctl");
-        return -2;
-      }
-      else
-      {
-        printf("evdev: Found '%s' at '%s'\n", name, device);
-
-        controller->fd = fd;
-
-        #if defined(TARGET_PANDORA)
-          *controller.mapping = &controller_mapping_pandora;
-        #elif defined(TARGET_GCW0)
-          *controller.mapping = &controller_mapping_gcwz;
-        #else
-          if (strcmp(name, "Microsoft X-Box 360 pad") == 0 ||
-              strcmp(name, "Xbox Gamepad (userspace driver)") == 0 ||
-              strcmp(name, "Xbox 360 Wireless Receiver (XBOX)") == 0)
-          {
-              controller->mapping = &controller_mapping_xbox360;
-          }
-          else
-          {
-              controller->mapping = &controller_mapping_generic;
-          }
-        #endif
-        printf("evdev: Using '%s' mapping\n", controller->mapping->name);
-
-        return 0;
-      }
-    }
-    else
-    {
-      perror("evdev: open");
-      return -1;
-    }
-  }
-
-  bool input_evdev_handle(Controller* controller, u32 port)
-  {
-    #define SET_FLAG(field, mask, expr) field =((expr) ? (field & ~mask) : (field | mask))
-    if (controller->fd < 0 || controller->mapping == NULL)
-    {
-      return false;
-    }
-
-    input_event ie;
-
-    while(read(controller->fd, &ie, sizeof(ie)) == sizeof(ie))
-    {
-      printf("type %i key %i state %i\n", ie.type, ie.code, ie.value);
-      switch(ie.type)
-      {
-        case EV_KEY:
-          if (ie.code == controller->mapping->Btn_A) {
-            SET_FLAG(kcode[port], DC_BTN_A, ie.value);
-          } else if (ie.code == controller->mapping->Btn_B) {
-            SET_FLAG(kcode[port], DC_BTN_B, ie.value);
-          } else if (ie.code == controller->mapping->Btn_C) {
-            SET_FLAG(kcode[port], DC_BTN_C, ie.value);
-          } else if (ie.code == controller->mapping->Btn_D) {
-            SET_FLAG(kcode[port], DC_BTN_D, ie.value);
-          } else if (ie.code == controller->mapping->Btn_X) {
-            SET_FLAG(kcode[port], DC_BTN_X, ie.value);
-          } else if (ie.code == controller->mapping->Btn_Y) {
-            SET_FLAG(kcode[port], DC_BTN_Y, ie.value);
-          } else if (ie.code == controller->mapping->Btn_Z) {
-            SET_FLAG(kcode[port], DC_BTN_Z, ie.value);
-          } else if (ie.code == controller->mapping->Btn_Start) {
-            SET_FLAG(kcode[port], DC_BTN_START, ie.value);
-          } else if (ie.code == controller->mapping->Btn_Escape) {
-            die("death by escape key");
-          } else if (ie.code == controller->mapping->Btn_DPad_Left) {
-            SET_FLAG(kcode[port], DC_DPAD_LEFT, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad_Right) {
-            SET_FLAG(kcode[port], DC_DPAD_RIGHT, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad_Up) {
-            SET_FLAG(kcode[port], DC_DPAD_UP, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad_Down) {
-            SET_FLAG(kcode[port], DC_DPAD_DOWN, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad2_Left) {
-            SET_FLAG(kcode[port], DC_DPAD2_LEFT, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad2_Right) {
-            SET_FLAG(kcode[port], DC_DPAD2_RIGHT, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad2_Up) {
-            SET_FLAG(kcode[port], DC_DPAD2_UP, ie.value);
-          } else if (ie.code == controller->mapping->Btn_DPad2_Down) {
-            SET_FLAG(kcode[port], DC_DPAD2_DOWN, ie.value);
-          } else if (ie.code == controller->mapping->Btn_Trigger_Left) {
-            lt[port] = (ie.value ? 255 : 0);
-          } else if (ie.code == controller->mapping->Btn_Trigger_Right) {
-            rt[port] = (ie.value ? 255 : 0);
-          }
-          break;
-        case EV_ABS:
-          if (ie.code == controller->mapping->Axis_DPad_X) {
-            switch(ie.value)
-            {
-              case -1:
-                  SET_FLAG(kcode[port], DC_DPAD_LEFT,  1);
-                  SET_FLAG(kcode[port], DC_DPAD_RIGHT, 0);
-                  break;
-              case 0:
-                  SET_FLAG(kcode[port], DC_DPAD_LEFT,  0);
-                  SET_FLAG(kcode[port], DC_DPAD_RIGHT, 0);
-                  break;
-              case 1:
-                  SET_FLAG(kcode[port], DC_DPAD_LEFT,  0);
-                  SET_FLAG(kcode[port], DC_DPAD_RIGHT, 1);
-                  break;
-            }
-          } else if (ie.code == controller->mapping->Axis_DPad_Y) {
-            switch(ie.value)
-            {
-              case -1:
-                  SET_FLAG(kcode[port], DC_DPAD_UP,   1);
-                  SET_FLAG(kcode[port], DC_DPAD_DOWN, 0);
-                  break;
-              case 0:
-                  SET_FLAG(kcode[port], DC_DPAD_UP,  0);
-                  SET_FLAG(kcode[port], DC_DPAD_DOWN, 0);
-                  break;
-              case 1:
-                  SET_FLAG(kcode[port], DC_DPAD_UP,  0);
-                  SET_FLAG(kcode[port], DC_DPAD_DOWN, 1);
-                  break;
-            }
-          } else if (ie.code == controller->mapping->Axis_DPad2_X) {
-            switch(ie.value)
-            {
-              case -1:
-                  SET_FLAG(kcode[port], DC_DPAD2_LEFT,  1);
-                  SET_FLAG(kcode[port], DC_DPAD2_RIGHT, 0);
-                  break;
-              case 0:
-                  SET_FLAG(kcode[port], DC_DPAD2_LEFT,  0);
-                  SET_FLAG(kcode[port], DC_DPAD2_RIGHT, 0);
-                  break;
-              case 1:
-                  SET_FLAG(kcode[port], DC_DPAD2_LEFT,  0);
-                  SET_FLAG(kcode[port], DC_DPAD2_RIGHT, 1);
-                  break;
-            }
-          } else if (ie.code == controller->mapping->Axis_DPad2_X) {
-            switch(ie.value)
-            {
-              case -1:
-                  SET_FLAG(kcode[port], DC_DPAD2_UP,   1);
-                  SET_FLAG(kcode[port], DC_DPAD2_DOWN, 0);
-                  break;
-              case 0:
-                  SET_FLAG(kcode[port], DC_DPAD2_UP,  0);
-                  SET_FLAG(kcode[port], DC_DPAD2_DOWN, 0);
-                  break;
-              case 1:
-                  SET_FLAG(kcode[port], DC_DPAD2_UP,  0);
-                  SET_FLAG(kcode[port], DC_DPAD2_DOWN, 1);
-                  break;
-            }
-          } else if (ie.code == controller->mapping->Axis_Analog_X) {
-            joyx[port] = (s8)(ie.value/256);
-          } else if (ie.code == controller->mapping->Axis_Analog_Y) {
-            joyy[port] = (s8)(ie.value/256);
-          } else if (ie.code == controller->mapping->Axis_Trigger_Left) {
-            lt[port] = (s8)ie.value;
-          } else if (ie.code == controller->mapping->Axis_Trigger_Right) {
-            rt[port] = (s8)ie.value;
-          }
-          break;
-      }
-    }
-  }
 #endif
 
-
-/* legacy joystick input */
-static int joystick_fd = -1; // Joystick file descriptor
-
 #if defined(USE_JOYSTICK)
-  #define JOYSTICK_DEVICE_STRING "/dev/input/js%d"
-  #define JOYSTICK_DEFAULT_DEVICE_ID 0
-  #define JOYSTICK_MAP_SIZE 32
-
-  const u32 joystick_map_btn_usb[JOYSTICK_MAP_SIZE]      = { DC_BTN_Y, DC_BTN_B, DC_BTN_A, DC_BTN_X, 0, 0, 0, 0, 0, DC_BTN_START };
-  const u32 joystick_map_axis_usb[JOYSTICK_MAP_SIZE]     = { DC_AXIS_X, DC_AXIS_Y, 0, 0, 0, 0, 0, 0, 0, 0 };
-
-  const u32 joystick_map_btn_xbox360[JOYSTICK_MAP_SIZE]  = { DC_BTN_A, DC_BTN_B, DC_BTN_X, DC_BTN_Y, 0, 0, 0, DC_BTN_START, 0, 0 };
-  const u32 joystick_map_axis_xbox360[JOYSTICK_MAP_SIZE] = { DC_AXIS_X, DC_AXIS_Y, DC_AXIS_LT, 0, 0, DC_AXIS_RT, DC_DPAD_LEFT, DC_DPAD_UP, 0, 0 };
-
-  const u32* joystick_map_btn = joystick_map_btn_usb;
-  const u32* joystick_map_axis = joystick_map_axis_usb;
-
-  int input_joystick_init(const char* device)
-  {
-    int axis_count = 0;
-    int button_count = 0;
-    char name[128] = "Unknown";
-
-    printf("joystick: Trying to open device at '%s'\n", device);
-
-    int fd = open(device, O_RDONLY);
-
-    if(fd >= 0)
-    {
-      fcntl(fd, F_SETFL, O_NONBLOCK);
-      ioctl(fd, JSIOCGAXES, &axis_count);
-      ioctl(fd, JSIOCGBUTTONS, &button_count);
-      ioctl(fd, JSIOCGNAME(sizeof(name)), &name);
-
-      printf("joystick: Found '%s' with %d axis and %d buttons at '%s'.\n", name, axis_count, button_count, device);
-
-      if (strcmp(name, "Microsoft X-Box 360 pad") == 0 ||
-          strcmp(name, "Xbox Gamepad (userspace driver)") == 0 ||
-          strcmp(name, "Xbox 360 Wireless Receiver (XBOX)") == 0)
-      {
-        joystick_map_btn = joystick_map_btn_xbox360;
-        joystick_map_axis = joystick_map_axis_xbox360;
-        printf("joystick: Using Xbox 360 map\n");
-      }
-    }
-    else
-    {
-      perror("joystick open");
-    }
-
-    return fd;
-  }
-
-  bool input_joystick_handle(int fd, u32 port)
-  {
-    // Joystick must be connected
-    if(fd < 0) {
-      return false;
-    }
-
-    struct js_event JE;
-    while(read(fd, &JE, sizeof(JE)) == sizeof(JE))
-    if (JE.number < JOYSTICK_MAP_SIZE)
-    {
-      switch(JE.type & ~JS_EVENT_INIT)
-      {
-        case JS_EVENT_AXIS:
-        {
-          u32 mt = joystick_map_axis[JE.number] >> 16;
-          u32 mo = joystick_map_axis[JE.number] & 0xFFFF;
-
-          //printf("AXIS %d,%d\n",JE.number,JE.value);
-          s8 v=(s8)(JE.value/256); //-127 ... + 127 range
-
-          if (mt == 0)
-          {
-            kcode[port] |= mo;
-            kcode[port] |= mo*2;
-            if (v<-64)
-            {
-              kcode[port] &= ~mo;
-            }
-            else if (v>64)
-            {
-              kcode[port] &= ~(mo*2);
-            }
-
-           //printf("Mapped to %d %d %d\n",mo,kcode[port]&mo,kcode[port]&(mo*2));
-          }
-          else if (mt == 1)
-          {
-            if (v >= 0)
-            {
-              v++;  //up to 255
-            }
-            //printf("AXIS %d,%d Mapped to %d %d %d\n",JE.number,JE.value,mo,v,v+127);
-            if (mo == 0)
-            {
-              lt[port] = (v + 127);
-            }
-            else if (mo == 1)
-            {
-              rt[port] = (v + 127);
-            }
-          }
-          else if (mt == 2)
-          {
-            //  printf("AXIS %d,%d Mapped to %d %d [%d]",JE.number,JE.value,mo,v);
-            if (mo == 0)
-            {
-              joyx[port] = v;
-            }
-            else if (mo == 1)
-            {
-              joyy[port] = v;
-            }
-          }
-        }
-        break;
-
-        case JS_EVENT_BUTTON:
-        {
-          u32 mt = joystick_map_btn[JE.number] >> 16;
-          u32 mo = joystick_map_btn[JE.number] & 0xFFFF;
-
-          // printf("BUTTON %d,%d\n",JE.number,JE.value);
-
-          if (mt == 0)
-          {
-            // printf("Mapped to %d\n",mo);
-            if (JE.value)
-            {
-              kcode[port] &= ~mo;
-            }
-            else
-            {
-              kcode[port] |= mo;
-            }
-          }
-          else if (mt == 1)
-          {
-            // printf("Mapped to %d %d\n",mo,JE.value?255:0);
-            if (mo==0)
-            {
-              lt[port] = JE.value ? 255 : 0;
-            }
-            else if (mo==1)
-            {
-              rt[port] = JE.value ? 255 : 0;
-            }
-          }
-        }
-        break;
-      }
-    }
-
-    return true;
-  }
+  /* legacy joystick input */
+  static int joystick_fd = -1; // Joystick file descriptor
 #endif
 
 void SetupInput()
