+++ core/linux-dist/main.h
@@ -1,0 +1,31 @@
+#pragma once
+
+extern u16 kcode[4];
+extern u32 vks[4];
+extern u8 rt[4], lt[4];
+extern s8 joyx[4], joyy[4];
+
+enum DreamcastController
+{
+  DC_BTN_C       = 1,
+  DC_BTN_B       = 1<<1,
+  DC_BTN_A       = 1<<2,
+  DC_BTN_START   = 1<<3,
+  DC_DPAD_UP     = 1<<4,
+  DC_DPAD_DOWN   = 1<<5,
+  DC_DPAD_LEFT   = 1<<6,
+  DC_DPAD_RIGHT  = 1<<7,
+  DC_BTN_Z       = 1<<8,
+  DC_BTN_Y       = 1<<9,
+  DC_BTN_X       = 1<<10,
+  DC_BTN_D       = 1<<11,
+  DC_DPAD2_UP    = 1<<12,
+  DC_DPAD2_DOWN  = 1<<13,
+  DC_DPAD2_LEFT  = 1<<14,
+  DC_DPAD2_RIGHT = 1<<15,
+
+  DC_AXIS_LT = 0X10000,
+  DC_AXIS_RT = 0X10001,
+  DC_AXIS_X  = 0X20000,
+  DC_AXIS_Y  = 0X20001,
+};
